backend/pyproject.toml
[project]
name = "ai-todo-backend"
version = "1.0.0"
description = "AI-Powered ToDo List Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.116.1",
    "uvicorn[standard]==0.35.0",
    "pydantic==2.11.7",
    "python-dotenv==1.1.1",
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-core>=1.0.0",
    "aiosqlite==0.19.0",
    "python-multipart==0.0.9",
    "requests==2.31.0",
    "asyncpg>=0.30.0",
    "sqlalchemy[asyncio]==2.0.23",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "langchain-postgres>=0.0.16",
    "langgraph-checkpoint-postgres>=3.0.0",
    "passlib[bcrypt]==1.7.4",
    "python-jose[cryptography]==3.3.0",
    "email-validator==2.1.0",
    "bcrypt==3.1.7",
    "assistant-stream==0.0.6",
    "weaviate-client==4.4.0",
    "openai>=1.109.1",
    "celery==5.3.4",
    "redis==5.0.1",
    "flower==2.0.1",
    "sentence-transformers==2.2.2",
    "numpy==1.24.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = []

